+++ qpsolvers/solvers/cvxopt_.py
@@ -16,6 +16,7 @@ using CVXOPT in a scientific work, consider citing the corresponding report
 **Warm-start:** this solver interface supports warm starting 🔥
 """
 
+import time
 import warnings
 from typing import Dict, Optional, Union
 
@@ -143,6 +144,7 @@ def cvxopt_solve_problem(
     parameters. See also [Caron2022]_ for a primer on the duality gap, primal
     and dual residuals.
     """
+    build_start_time = time.perf_counter()
     P, q, G, h, A, b, lb, ub = problem.unpack()
     if lb is not None or ub is not None:
         G, h = linear_from_box_inequalities(
@@ -165,6 +167,7 @@ def cvxopt_solve_problem(
     kwargs["show_progress"] = verbose
 
     try:
+        solve_start_time = time.perf_counter()
         res = qp(
             *args,
             solver=solver,
@@ -172,6 +175,7 @@ def cvxopt_solve_problem(
             options=kwargs,
             **constraints,
         )
+        solve_end_time = time.perf_counter()
     except ValueError as exception:
         error = str(exception)
         if "Rank(A)" in error:
@@ -195,6 +199,8 @@ def cvxopt_solve_problem(
         solution.z = np.empty((0,))
         solution.z_box = np.empty((0,))
     solution.obj = res["primal objective"]
+    solution.build_time = solve_start_time - build_start_time
+    solution.solve_time = solve_end_time - solve_start_time
     return solution