+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add build and solve timing info (thanks to @ahoarau)
 - New solver: qpmad (thanks to @ahoarau)
 - Support Python 3.13